app/django_api/requirements.txt
opencv-python
redis
pika
requests
djangorestframework
drf-yasg
django
psycopg2